# apps/decision-log/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Decision Log API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Person {
  id: string;
  name: string;
  initials: string;
}

export type ReasonRepeatability = typeof ReasonRepeatability[keyof typeof ReasonRepeatability];


export const ReasonRepeatability = {
  safe: 'safe',
  sensitive: 'sensitive',
  unknown: 'unknown',
} as const;

export type ReasonState = typeof ReasonState[keyof typeof ReasonState];


export const ReasonState = {
  current: 'current',
  superseded: 'superseded',
} as const;

export interface Reason {
  id: string;
  text: string;
  author: Person;
  createdAt: string;
  repeatability: ReasonRepeatability;
  state: ReasonState;
  permalink: string;
}

export type DecisionChannelType = typeof DecisionChannelType[keyof typeof DecisionChannelType];


export const DecisionChannelType = {
  public: 'public',
  private: 'private',
} as const;

export type DecisionStatus = typeof DecisionStatus[keyof typeof DecisionStatus];


export const DecisionStatus = {
  current: 'current',
  reversed: 'reversed',
} as const;

export type DecisionSourceState = typeof DecisionSourceState[keyof typeof DecisionSourceState];


export const DecisionSourceState = {
  active: 'active',
  deleted: 'deleted',
} as const;

export interface Decision {
  id: string;
  title: string;
  text: string;
  channel: string;
  channelType?: DecisionChannelType;
  status: DecisionStatus;
  sourceState: DecisionSourceState;
  /** @nullable */
  sourceEditedAt?: string | null;
  decidedBy: Person;
  loggedBy: Person;
  decidedAt: string;
  permalink: string;
  /** @nullable */
  replacesId?: string | null;
  /** @nullable */
  replacementTitle?: string | null;
  reasons: Reason[];
  participantCount: number;
  askCount?: number;
}

export type DecisionInputChannelType = typeof DecisionInputChannelType[keyof typeof DecisionInputChannelType];


export const DecisionInputChannelType = {
  public: 'public',
  private: 'private',
} as const;

export interface DecisionInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  text: string;
  /** @minLength 1 */
  channel: string;
  channelType: DecisionInputChannelType;
  permalink: string;
  decidedById: string;
  loggedById: string;
}

export type ReasonInputRepeatability = typeof ReasonInputRepeatability[keyof typeof ReasonInputRepeatability];


export const ReasonInputRepeatability = {
  safe: 'safe',
  sensitive: 'sensitive',
  unknown: 'unknown',
} as const;

export interface ReasonInput {
  /** @minLength 1 */
  text: string;
  authorId: string;
  repeatability?: ReasonInputRepeatability;
}

export interface WhyInput {
  /** @minLength 1 */
  query: string;
  askerId?: string;
}

export type WhyAnswerKind = typeof WhyAnswerKind[keyof typeof WhyAnswerKind];


export const WhyAnswerKind = {
  answer: 'answer',
  no_match: 'no_match',
  tombstone: 'tombstone',
} as const;

export interface WhyAnswer {
  kind: WhyAnswerKind;
  query: string;
  message: string;
  decision?: Decision | null;
  matches: Decision[];
  canAsk?: boolean;
}

export interface Summary {
  totalDecisions: number;
  withReasons: number;
  missingReasons: number;
  currentDecisions: number;
  captureRate: number;
  averageTimeToReasonMinutes: number;
  activeChannels: number;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  decision_logged: 'decision_logged',
  reason_captured: 'reason_captured',
  reason_replaced: 'reason_replaced',
  source_edited: 'source_edited',
  source_deleted: 'source_deleted',
  ask_sent: 'ask_sent',
} as const;

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  label: string;
  detail: string;
  createdAt: string;
  person: Person;
  /** @nullable */
  decisionId?: string | null;
}

export interface ActionResult {
  success: boolean;
  message: string;
}

export type ListActivityParams = {
/**
 * @minimum 1
 * @maximum 20
 */
limit?: number;
};

export type ListDecisionsParams = {
query?: string;
status?: ListDecisionsStatus;
channel?: string;
};

export type ListDecisionsStatus = typeof ListDecisionsStatus[keyof typeof ListDecisionsStatus];


export const ListDecisionsStatus = {
  all: 'all',
  current: 'current',
  reversed: 'reversed',
  incomplete: 'incomplete',
} as const;