# apps/decision-log/lib/api-zod/src/generated/types/decisionInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Decision Log API
 * OpenAPI spec version: 0.1.0
 */
import type { DecisionInputChannelType } from './decisionInputChannelType';

export interface DecisionInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  text: string;
  /** @minLength 1 */
  channel: string;
  channelType: DecisionInputChannelType;
  permalink: string;
  decidedById: string;
  loggedById: string;
}