# apps/decision-log/lib/api-zod/src/generated/types/reasonInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Decision Log API
 * OpenAPI spec version: 0.1.0
 */
import type { ReasonInputRepeatability } from './reasonInputRepeatability';

export interface ReasonInput {
  /** @minLength 1 */
  text: string;
  authorId: string;
  repeatability?: ReasonInputRepeatability;
}